+++ CHANGELOG.md
@@ -15,10 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - When a task is validated
   - When a task is upcoming (2 weeks, 3 days ahead and on the day of the task)
   - When a task is overdue (only 1 week after for shifts)
+  - When a task is updated and a notification is requested
 
 ### Changed
 
+- Allow editors to change timing of helper tasks
+- Do not allow changing the year of the task if anyone has signed up
 - Hierarchical `config.json` for better readability
+- UTF-8 log file
 
 ## [1.1.0] - 2025-03-16